long

+++ src/Application.Dtos/OrderDetailDto.cs
@@ -3,10 +3,10 @@ namespace Application.Dtos;
 [Serializable]
 public class OrderDetailDto
 {
-    public int Id { get; set; }
-    public int OrderId { get; set; }
-    public int ProductId { get; set; }
-    public int OrderQty { get; set; }
+    public long Id { get; set; }
+    public long OrderId { get; set; }
+    public long ProductId { get; set; }
+    public long OrderQty { get; set; }
     public decimal UnitPrice { get; set; }
     public decimal Total { get; set; }
 }
