+++ src/Domain/Entities/Product.cs
@@ -5,5 +5,5 @@ public class Product
     public long Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public decimal Price { get; set; }
-    public int Stock { get; set; }
+    public long Stock { get; set; }
 }
